--- apps/web/src/components/location-setting/index.ts
@@ -1,4 +1,0 @@
-export { default as LocationList } from './location-list';
-export { default as LocationItem } from './location-item';
-export { default as ButtonBox } from './button-box';
-export { default as ButtonCreate } from './button-create';
